+++ CHANGELOG.md
@@ -236,6 +236,8 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Updated `player` to `0.5.1` from `0.6.1`.
 - Updated streamchild render method to use default behavior when using default blocks [Fixed 1268] (https://github.com/cfpb/cfgov-refresh/issues/1268)
 - Fixes styling and rendering issues [Fixed 1278] (https://github.com/cfpb/cfgov-refresh/issues/1278)
+- Upgrade version of Wagtail to 1.3
+- Change method of CFGOVPage called `children` to be called `elements`
 
 ### Removed
 - Removed unused exportsOverride section,
@@ -260,6 +262,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Fix related post molecule to be used in multiple places
 - Fix failing tests relating to Related Posts organism
 - Fix related-posts.html logic
+- Minor PEP8 compliance changes
 
 
 ## 3.0.0-2.4.0 - 2015-09-29